Add transformApiEvent method to helper
