+++ frontend/src/utils/transform-event-data.js
@@ -1,4 +1,4 @@
-const transformEventData = (inputValues) => {
+export const transformNewEventData = (inputValues) => {
   const {
     title,
     summary,
@@ -31,4 +31,25 @@ const transformEventData = (inputValues) => {
   return event;
 };
 
-export default transformEventData;
+export const transformApiEvent = (event) => {
+  const eventObj = {
+    id: event.id,
+    title: event.name.text,
+    description: event.description.text,
+    date: event.start.local,
+    location: event.venue?.name ? event.venue.name : 'Online',
+    address: event.venue?.address?.localized_address_display
+      ? event.venue.address.localized_address_display
+      : '',
+    imageOriginal: event.logo.original.url,
+    image: event.logo.url,
+    category: event.category.name,
+    availableTickets: event.capacity,
+    price: event.ticket_availability.maximum_ticket_price.value,
+    displayPrice: event.ticket_availability.maximum_ticket_price.major_value,
+    currency: event.ticket_availability.maximum_ticket_price.currency,
+    ticketsWishList: 0,
+  };
+
+  return eventObj;
+};
